refactor: data와 domain 레이어 분리 및 repository interface로 전환
- data와 domain 레이어를 분리하여 아키텍처를 개선했습니다.
- Repository 인터페이스를 domain 레이어로 이동시켰습니다.
- 기존 Repository 구현체를 `Default` 접두사를 붙여 data 레이어에 두었습니다.
- Model 패키지를 domain 레이어 하위로 이동했습니다.
- 이에 따라 관련 클래스와 XML 파일의 import 경로를 수정했습니다.
