+++ app/src/main/java/woowacourse/shopping/data/repository/DefaultProductRepository.kt
@@ -1,8 +1,9 @@
-package woowacourse.shopping.data
+package woowacourse.shopping.data.repository
 
-import woowacourse.shopping.model.Product
+import woowacourse.shopping.domain.model.Product
+import woowacourse.shopping.domain.repository.ProductRepository
 
-class ProductRepository {
+class DefaultProductRepository : ProductRepository {
     private val products: List<Product> =
         listOf(
             Product(
@@ -22,5 +23,5 @@ class ProductRepository {
             ),
         )
 
-    fun getAllProducts(): List<Product> = products
+    override fun getAllProducts(): List<Product> = products
 }
