+++ app/src/main/java/woowacourse/shopping/di/AppContainer.kt
@@ -1,14 +1,16 @@
 package woowacourse.shopping.di
 
-import woowacourse.shopping.data.CartRepository
-import woowacourse.shopping.data.ProductRepository
+import woowacourse.shopping.data.repository.DefaultCartRepository
+import woowacourse.shopping.data.repository.DefaultProductRepository
+import woowacourse.shopping.domain.repository.CartRepository
+import woowacourse.shopping.domain.repository.ProductRepository
 
 class AppContainer {
     val cartRepository: CartRepository by lazy {
-        CartRepository()
+        DefaultCartRepository()
     }
 
     val productRepository: ProductRepository by lazy {
-        ProductRepository()
+        DefaultProductRepository()
     }
 }
