+++ app/src/main/java/woowacourse/shopping/domain/repository/CartRepository.kt
@@ -1,0 +1,11 @@
+package woowacourse.shopping.domain.repository
+
+import woowacourse.shopping.domain.model.Product
+
+interface CartRepository {
+    fun addCartProduct(product: Product)
+
+    fun getAllCartProducts(): List<Product>
+
+    fun deleteCartProduct(id: Int)
+}
