+++ app/src/main/java/woowacourse/shopping/ui/cart/CartProductAdapter.kt
@@ -2,7 +2,7 @@ package woowacourse.shopping.ui.cart
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import woowacourse.shopping.model.Product
+import woowacourse.shopping.domain.model.Product
 
 class CartProductAdapter(
     items: List<Product>,
