+++ app/src/main/java/woowacourse/shopping/ui/cart/CartViewModel.kt
@@ -3,11 +3,11 @@ package woowacourse.shopping.ui.cart
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import woowacourse.shopping.data.CartRepository
-import woowacourse.shopping.model.Product
+import woowacourse.shopping.data.repository.DefaultCartRepository
+import woowacourse.shopping.domain.model.Product
 
 class CartViewModel(
-    private val cartRepository: CartRepository,
+    private val cartRepository: DefaultCartRepository,
 ) : ViewModel() {
     private val _cartProducts: MutableLiveData<List<Product>> =
         MutableLiveData(emptyList())
